otel-news-service/requirements.txt
# For OpenTelemetry SDK and API
opentelemetry-sdk

# For AWS Interaction
boto3
